requirements.txt
pafy
python-vlc
rumps
